Move TreeArrayView to separate file
