+++ aim/storage/inmemorytreeview.py
@@ -2,7 +2,7 @@ from copy import deepcopy
 
 from aim.storage.types import CustomObjectBase
 from aim.storage.types import AimObject, AimObjectKey, AimObjectPath
-from aim.storage.arrayview import TreeArrayView
+from aim.storage.treearrayview import TreeArrayView
 from aim.storage.treeview import TreeView
 
 from typing import Iterator, Tuple, Union
