+++ aim/storage/treeviewproxy.py
@@ -3,7 +3,7 @@ import struct
 from aim.ext.transport.message_utils import ResourceObject
 from aim.ext.transport.remote_resource import RemoteResourceAutoClean
 from aim.storage.treeview import TreeView
-from aim.storage.arrayview import TreeArrayView
+from aim.storage.treearrayview import TreeArrayView
 from aim.storage.types import AimObject, AimObjectKey, AimObjectPath
 
 from typing import TYPE_CHECKING, Iterator, Tuple, Union
